% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_0s_by_col.R
\name{rm_0s_by_col}
\alias{rm_0s_by_col}
\title{Remove columns with all 0s in any given matrix}
\usage{
rm_0s_by_col(mat)
}
\arguments{
\item{mat}{A numeric matrix.}
}
\value{
A matrix with columns removed if all elements in the column are 0.
}
\description{
This function removes columns from a matrix where all elements are 0.
}
\examples{
x <- matrix(rpois(100, 0.1), 10, 10)
rm_0s_by_col(x)
}
