% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_0s_by_row.R
\name{rm_0s_by_row}
\alias{rm_0s_by_row}
\title{Remove rows with all 0s in any given matrix}
\usage{
rm_0s_by_row(mat)
}
\arguments{
\item{mat}{A numeric matrix.}
}
\value{
A matrix with rows removed if all elements in the row are 0.
}
\description{
This function removes rows from a matrix where all elements are 0.
}
\examples{
x <- matrix(rpois(100, 0.1), 10, 10)
rm_0s_by_row(x)
}
